requirements.txt
# Core Dependencies
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0

# System Monitoring
psutil>=5.9.0

# Configuration and Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
rich>=13.0.0
chardet>=5.2.0

# HTTP and Networking
aiohttp>=3.8.0
requests>=2.31.0

# Database
# sqlite3 is built-in

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Enhanced Features
cryptography>=41.0.0

# Optional Dependencies
# Uncomment for additional features
# openai>=1.0.0
# langchain>=0.1.0
# transformers>=4.30.0 